refactor: モックのデータと通信を別のフォルダに移動
