+++ mocks/data/categories.ts
@@ -1,0 +1,33 @@
+export const categories = [
+  {
+    id: 'oeijfeowijfwoeijk',
+    name: 'Work',
+  },
+  {
+    id: 'feoijfeifjeoifjk',
+    name: 'お買物リスト',
+  },
+  {
+    id: 'feoijfeoijfeoifj',
+    name: '買いたい',
+  },
+  {
+    id: 'fwoefjwlekffewogn',
+    name: 'House',
+  },
+  {
+    id: 'rigjrokanklrkwgnk',
+    name: 'その他',
+  },
+];
+
+export const manyCategory = [
+  ...categories,
+  { id: 'awpfwkafewlk', name: 'その他' },
+  { id: 'fefefewfewaf', name: 'その他' },
+  { id: 'thegrgragrgr', name: 'その他' },
+  { id: 'bebebbgbtrte', name: 'その他' },
+  { id: 'ebtbgbrbrtbr', name: 'その他' },
+  { id: 'oijkmpokekll', name: 'その他' },
+  { id: 'pkmjwemrkmkw', name: 'その他' },
+];
