+++ mocks/data/index.ts
@@ -1,0 +1,11 @@
+import { categories, manyCategory } from './categories';
+import { task, tasksSortWithCategory, tasksSortWithCreatedAt, tasksSortWithLimit } from './tasks';
+
+export {
+  categories,
+  manyCategory,
+  task,
+  tasksSortWithCategory,
+  tasksSortWithCreatedAt,
+  tasksSortWithLimit,
+};
