+++ mocks/handlers/index.ts
@@ -1,0 +1,21 @@
+import { updateCategoryHeader } from './categories';
+import {
+  createTasksErrorHeader,
+  createTasksHeader,
+  successDeleteTaskHandler,
+  successTaskAddHandler,
+  successUpdateTaskHandler,
+  updateTasksErrorHeader,
+  updateTasksHeader,
+} from './tasks';
+
+export {
+  updateCategoryHeader,
+  createTasksHeader,
+  createTasksErrorHeader,
+  updateTasksHeader,
+  updateTasksErrorHeader,
+  successUpdateTaskHandler,
+  successDeleteTaskHandler,
+  successTaskAddHandler,
+};
