+++ stories/model/task/EditTaskForm.stories.tsx
@@ -1,9 +1,11 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { rest } from 'msw';
 import React from 'react';
 import { Middleware, SWRConfig, SWRResponse } from 'swr';
 
 import { EditTaskForm } from '../../../components/model/task/EditTaskForm';
+import { categories } from '../../../mocks/data';
+import { task } from '../../../mocks/data/tasks';
+import { updateTasksErrorHeader, updateTasksHeader } from '../../../mocks/handlers';
 
 const meta: ComponentMeta<typeof EditTaskForm> = {
   component: EditTaskForm,
@@ -12,36 +14,6 @@ const meta: ComponentMeta<typeof EditTaskForm> = {
 export default meta;
 
 const Template: ComponentStory<typeof EditTaskForm> = (args) => <EditTaskForm {...args} />;
-
-const categories = [
-  {
-    id: 'oeijfeowijfwoeijk',
-    name: 'Work',
-  },
-  {
-    id: 'feoijfeifjeoifjk',
-    name: 'お買物リスト',
-  },
-  {
-    id: 'feoijfeoijfeoifj',
-    name: '買いたい',
-  },
-  {
-    id: 'fwoefjwlekffewogn',
-    name: 'House',
-  },
-  {
-    id: 'rigjrokanklrkwgnk',
-    name: 'その他',
-  },
-];
-
-const task = {
-  title: '{"title":"お米を買う","checked":false,"detail":"次はあきたこまちがいいかもしれない"}',
-  category: '買い物リスト',
-  limit: '2022-03-05',
-  detail: '',
-};
 
 const middleware: Middleware = () => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -83,21 +55,7 @@ Default.args = {
 Default.decorators = [(story) => <SWRConfig value={{ use: [middleware] }}>{story()}</SWRConfig>];
 Default.parameters = {
   msw: {
-    handlers: [
-      rest.patch(`${process.env.NEXT_PUBLIC_API_ORIGIN}/api/tasks/pirgnojgn`, (req, res, ctx) => {
-        return res(
-          ctx.status(200),
-          ctx.json({
-            id: 'aJDm3esbPPlGLKseXPXp',
-            title: 'お米を買う',
-            category: '買い物リスト',
-            limit: '2020/5/4',
-            detail: '次はあきたこまちがいいかもしれない',
-            created_at: '2020-03-28T07:55:21.109Z',
-          })
-        );
-      }),
-    ],
+    handlers: [updateTasksHeader],
   },
 };
 
@@ -116,15 +74,6 @@ HttpError.args = {
 HttpError.decorators = [(story) => <SWRConfig value={{ use: [middleware] }}>{story()}</SWRConfig>];
 HttpError.parameters = {
   msw: {
-    handlers: [
-      rest.patch(`${process.env.NEXT_PUBLIC_API_ORIGIN}/api/taskspirgnojgn`, (req, res, ctx) => {
-        return res(
-          ctx.status(401),
-          ctx.json({
-            error: 'Unauthorized',
-          })
-        );
-      }),
-    ],
+    handlers: [updateTasksErrorHeader],
   },
 };
