+++ stories/model/task/TaskList.stories.tsx
@@ -1,8 +1,19 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { rest } from 'msw';
 import { Middleware, SWRConfig, SWRResponse } from 'swr';
 
 import { TaskList } from '../../../components/model/task/TaskList';
+import {
+  categories,
+  task,
+  tasksSortWithCategory,
+  tasksSortWithCreatedAt,
+  tasksSortWithLimit,
+} from '../../../mocks/data';
+import {
+  successDeleteTaskHandler,
+  successTaskAddHandler,
+  successUpdateTaskHandler,
+} from '../../../mocks/handlers';
 
 const meta: ComponentMeta<typeof TaskList> = {
   component: TaskList,
@@ -11,74 +22,6 @@ const meta: ComponentMeta<typeof TaskList> = {
 export default meta;
 
 const Template: ComponentStory<typeof TaskList> = (args) => <TaskList {...args} />;
-
-const tasks = [
-  {
-    id: 'aJDm3esbPPlGLKseXPXp',
-    title: '{"title":"お米を買う","checked":false,"detail":"次はあきたこまちがいいかもしれない"}',
-    category: '買い物リスト',
-    limit: '2020-5-4',
-    created_at: '2020-03-28T07:55:21.109Z',
-  },
-  {
-    id: 'zRTxMgbCsPCKsbUyU5cg',
-    title:
-      '{"title":"高沼プロジェクトを終わらせる","checked":false,"detail":"次はあきたこまちがいいかもしれない"}',
-    category: 'Work',
-    limit: '2020-6-11',
-    created_at: '2020-03-27T08:20:11.109Z',
-  },
-  {
-    id: 'zRTxMgbCsPCKsbUyU5c1',
-    title:
-      '{"title":"GraphQLについて勉強する","checked":true,"detail":"次はあきたこまちがいいかもしれない"}',
-    category: 'GraphQL',
-    limit: '2020-7-15',
-    created_at: '2020-03-26T08:20:11.109Z',
-  },
-  {
-    id: 'zRTxMgbCsPCKsbUyU5c2',
-    title:
-      '{"title":"ESLintについて勉強する","checked":true,"detail":"次はあきたこまちがいいかもしれない"}',
-    category: 'Work',
-    limit: '2020-8-23',
-    created_at: '2020-03-25T08:20:11.109Z',
-  },
-] as const;
-
-const tasksSortWithCategory = [tasks[2], tasks[1], tasks[3], tasks[0]];
-const tasksSortWithCreatedAt = [tasks[3], tasks[2], tasks[1], tasks[0]];
-const tasksSortWithLimit = [tasks[0], tasks[1], tasks[2], tasks[3]];
-
-const task = {
-  title: '{"title":"お米を買う","checked":false,"detail":"次はあきたこまちがいいかもしれない"}',
-  category: '買い物リスト',
-  limit: '2022-03-05',
-  detail: '',
-};
-
-const categories = [
-  {
-    id: 'oeijfeowijfwoeijk',
-    name: 'Work',
-  },
-  {
-    id: 'feoijfeifjeoifjk',
-    name: 'お買物リスト',
-  },
-  {
-    id: 'feoijfeoijfeoifj',
-    name: '買いたい',
-  },
-  {
-    id: 'fwoefjwlekffewogn',
-    name: 'House',
-  },
-  {
-    id: 'rigjrokanklrkwgnk',
-    name: 'その他',
-  },
-];
 
 const middleware: Middleware = () => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -89,7 +32,6 @@ const middleware: Middleware = () => {
     }
     const mockData: { [name: string]: unknown } = {
       '/api/tasks/aJDm3esbPPlGLKseXPXp': task,
-      '/api/tasks': tasks,
       '/api/tasks?sort=created_at': tasksSortWithCreatedAt,
       '/api/tasks?sort=category': tasksSortWithCategory,
       '/api/tasks?sort=limit': tasksSortWithLimit,
@@ -117,49 +59,6 @@ const loadingMiddleware: Middleware = () => {
   };
 };
 
-const successDeleteHandler = rest.delete(
-  `${process.env.NEXT_PUBLIC_API_ORIGIN}/api/tasks/*`,
-  (req, res, ctx) => {
-    return res(
-      ctx.json({
-        message: 'Successfully deleted task.',
-      })
-    );
-  }
-);
-
-const successUpdateHandler = rest.patch(
-  `${process.env.NEXT_PUBLIC_API_ORIGIN}/api/tasks/*`,
-  (req, res, ctx) => {
-    return res(
-      ctx.json({
-        title:
-          '{"title":"高沼プロジェクトを終わらせる","checked":false,"detail":"次はあきたこまちがいいかもしれない"}',
-        category: '買い物リスト',
-        limit: '2020/5/4',
-        detail: '',
-      })
-    );
-  }
-);
-
-const successTaskAddHandler = rest.post(
-  `${process.env.NEXT_PUBLIC_API_ORIGIN}/api/tasks`,
-  (req, res, ctx) => {
-    return res(
-      ctx.status(200),
-      ctx.json({
-        id: 'aJDm3esbPPlGLKseXPXp',
-        title: 'お米を買う',
-        category: '買い物リスト',
-        limit: '2020/5/4',
-        detail: '次はあきたこまちがいいかもしれない',
-        created_at: '2020-03-28T07:55:21.109Z',
-      })
-    );
-  }
-);
-
 export const AllCategories = Template.bind({});
 AllCategories.args = {
   category: '',
@@ -169,7 +68,7 @@ AllCategories.decorators = [
 ];
 AllCategories.parameters = {
   msw: {
-    handlers: [successDeleteHandler, successUpdateHandler, successTaskAddHandler],
+    handlers: [successDeleteTaskHandler, successUpdateTaskHandler, successTaskAddHandler],
   },
 };
 
@@ -190,6 +89,6 @@ SelectCategory.decorators = [
 ];
 SelectCategory.parameters = {
   msw: {
-    handlers: [successDeleteHandler, successUpdateHandler],
+    handlers: [successDeleteTaskHandler, successUpdateTaskHandler],
   },
 };
