+++ stories/provider/ModalProvider.stories.tsx
@@ -5,6 +5,7 @@ import { AddTaskForm } from '../../components/model/task/AddTaskForm';
 import { EditTaskForm } from '../../components/model/task/EditTaskForm';
 import { ModalProvider } from '../../components/provider/ModalProvider';
 import { Alert } from '../../components/ui/utils/Alert';
+import { categories, task } from '../../mocks/data';
 
 const meta: ComponentMeta<typeof ModalProvider> = {
   component: ModalProvider,
@@ -19,36 +20,6 @@ const meta: ComponentMeta<typeof ModalProvider> = {
 export default meta;
 
 const Template: ComponentStory<typeof ModalProvider> = (args) => <ModalProvider {...args} />;
-
-const categories = [
-  {
-    id: 'oeijfeowijfwoeijk',
-    name: 'Work',
-  },
-  {
-    id: 'feoijfeifjeoifjk',
-    name: 'お買物リスト',
-  },
-  {
-    id: 'feoijfeoijfeoifj',
-    name: '買いたい',
-  },
-  {
-    id: 'fwoefjwlekffewogn',
-    name: 'House',
-  },
-  {
-    id: 'rigjrokanklrkwgnk',
-    name: 'その他',
-  },
-];
-
-const task = {
-  title: '{"title":"お米を買う","checked":false,"detail":"次はあきたこまちがいいかもしれない"}',
-  category: '買い物リスト',
-  limit: '2022-03-05',
-  detail: '',
-};
 
 const middleware: Middleware = () => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
